requirements.txt
pandas
requests
requests-html
requests-toolbelt
numpy
bs4
apache-airflow
PyGithub
pymongo
overpy
python-dotenv